pray_across_europe/requirements.txt
asgiref==3.5.2
awsebcli==3.20.3
botocore==1.23.54
cement==2.8.2
certifi==2022.9.24
charset-normalizer==2.0.12
colorama==0.4.3
Django==4.0.6
django-crispy-forms==1.14.0
future==0.16.0
idna==3.4
jmespath==0.10.0
npm==0.1.1
optional-django==0.1.0
pathspec==0.9.0
Pillow==9.2.0
python-dateutil==2.8.2
python-dotenv==0.20.0
PyYAML==5.4.1
requests==2.26.0
six==1.14.0
sqlparse==0.4.2
termcolor==1.1.0
tzdata==2022.1
urllib3==1.26.13
wcwidth==0.1.9
